requirements.txt
fastapi
uvicorn
qrcode
Pillow
sqlmodel
psycopg2-binary
python-dotenv
alembicpasslib[bcrypt]
python-jose
httpx==0.27.0
hyperframe<6.0.0
h2<5.0.0
